Changed add, update and delete requests to use proper request types and use standard REST urls, Changed add and update to use JSON data.

+++ src/main/java/janisl/musicdb/controllers/LabelController.java
@@ -2,19 +2,22 @@ package janisl.musicdb.controllers;
 
 import janisl.musicdb.HibernateUtil;
 import janisl.musicdb.models.Label;
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.util.List;
 import org.hibernate.Session;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping(value = "/label")
 public class LabelController {
 
-    @RequestMapping(value = "/list", method = RequestMethod.GET)
+    @RequestMapping(value = "/", method = RequestMethod.GET)
     public List<Label> getList() {
         Session session = HibernateUtil.getSessionFactory().openSession();
         session.beginTransaction();
@@ -23,10 +26,9 @@ public class LabelController {
 
         session.close();
         return labels;
-
     }
 
-    @RequestMapping(value = "/{id}")
+    @RequestMapping(value = "/{id}", method = RequestMethod.GET)
     public Label getById(@PathVariable("id") int id) {
         Session session = HibernateUtil.getSessionFactory().openSession();
         session.beginTransaction();
@@ -37,58 +39,55 @@ public class LabelController {
         return label;
     }
 
-    @RequestMapping(value = "/add")
-    public Label add(
-            @RequestParam(value = "name") String name,
-            @RequestParam(value = "beatportId", required = false) Integer beatportId,
-            @RequestParam(value = "beatportUrl", required = false) String beatportUrl,
-            @RequestParam(value = "discogsId", required = false) Integer discogsId) {
+    @RequestMapping(value = "/", method = RequestMethod.POST)
+    public ResponseEntity add(@RequestBody Label label) {
         Session session = HibernateUtil.getSessionFactory().openSession();
         session.beginTransaction();
-
-        Label label = new Label();
-        label.setName(name);
-        label.setBeatportId(beatportId);
-        label.setBeatportUrl(beatportUrl);
-        label.setDiscogsId(discogsId);
 
         session.save(label);
         session.getTransaction().commit();
         session.close();
-        return label;
+        URI locationUri;
+        try {
+            locationUri = new URI("/label/" + label.getId().toString());
+        } catch (URISyntaxException e) {
+            return ResponseEntity.badRequest().build();
+        }
+        return ResponseEntity.ok().location(locationUri).build();
     }
 
-    @RequestMapping(value = "/update/{id}")
-    public void update(
-            @PathVariable("id") int id,
-            @RequestParam(value = "name") String name,
-            @RequestParam(value = "beatportId") Integer beatportId,
-            @RequestParam(value = "beatportUrl") String beatportUrl,
-            @RequestParam(value = "discogsId") Integer discogsId) {
+    @RequestMapping(value = "/{id}", method = RequestMethod.PUT)
+    public ResponseEntity update(@PathVariable("id") int id, @RequestBody Label newLabel) {
         Session session = HibernateUtil.getSessionFactory().openSession();
         session.beginTransaction();
 
         Label label = (Label) session.get(Label.class, id);
-        label.setName(name);
-        label.setBeatportId(beatportId);
-        label.setBeatportUrl(beatportUrl);
-        label.setDiscogsId(discogsId);
+        if (label == null)
+            return ResponseEntity.notFound().build();
+        label.setName(newLabel.getName());
+        label.setBeatportId(newLabel.getBeatportId());
+        label.setBeatportUrl(newLabel.getBeatportUrl());
+        label.setDiscogsId(newLabel.getDiscogsId());
 
         session.save(label);
         session.getTransaction().commit();
         session.close();
+        return ResponseEntity.ok().build();
     }
 
-    @RequestMapping(value = "/delete/{id}")
-    public void delete(@PathVariable("id") int id) {
+    @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
+    public ResponseEntity delete(@PathVariable("id") int id) {
         Session session = HibernateUtil.getSessionFactory().openSession();
         session.beginTransaction();
 
         Label label = (Label) session.get(Label.class, id);
+        if (label == null)
+            return ResponseEntity.notFound().build();
 
         session.delete(label);
         session.getTransaction().commit();
         session.close();
+        return ResponseEntity.ok().build();
     }
 
 }
